pixel2: New system image changes

+++ App/USBStorageMode/launch.sh
@@ -29,7 +29,7 @@ case "$PLATFORM" in
         ;;
     "Pixel2")
         STORAGE_DEVICE="/dev/mmcblk0p3"
-        MOUNT_POINT="/storage/games-external"
+        MOUNT_POINT="/mnt/SDCARD/"
         USB_GADGET_PATH="/sys/kernel/config/usb_gadget/rockchip"
         USB_UDC_CONTROLLER="ff300000.usb"
         USB_CONFIG_PATH="$USB_GADGET_PATH/configs/b.1"
@@ -253,10 +253,15 @@ while true; do
         sleep 3
         stop_pyui_message_writer
         sync
-        cp "$STAGE_2_PATH" "$STAGE_2_TMP" && chmod +x "$STAGE_2_TMP"
-        export PATH=/usr/bin:/usr/sbin:/bin:/sbin
-        unset LD_LIBRARY_PATH
-        exec "$STAGE_2_TMP" --reboot
+
+        if device_system_handles_sdcard_unmount; then
+            device_run_reboot_cmd
+        else
+            cp "$STAGE_2_PATH" "$STAGE_2_TMP" && chmod +x "$STAGE_2_TMP"
+            export PATH=/usr/bin:/usr/sbin:/bin:/sbin
+            unset LD_LIBRARY_PATH
+            exec "$STAGE_2_TMP" --reboot
+        fi
     fi
 
     log_and_display_message "USB Mode Active.\nPress A to exit and reboot your device."
@@ -266,10 +271,15 @@ while true; do
         sleep 3
         stop_pyui_message_writer
         sync
-        cp "$STAGE_2_PATH" "$STAGE_2_TMP" && chmod +x "$STAGE_2_TMP"
-        export PATH=/usr/bin:/usr/sbin:/bin:/sbin
-        unset LD_LIBRARY_PATH
-        exec "$STAGE_2_TMP" --reboot
+
+        if device_system_handles_sdcard_unmount; then
+            device_run_reboot_cmd
+        else
+            cp "$STAGE_2_PATH" "$STAGE_2_TMP" && chmod +x "$STAGE_2_TMP"
+            export PATH=/usr/bin:/usr/sbin:/bin:/sbin
+            unset LD_LIBRARY_PATH
+            exec "$STAGE_2_TMP" --reboot
+        fi
     fi
     # Add a small sleep to prevent the loop from overwhelming the CPU
     sleep 1
